server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fuzzysearch==0.7.2",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.1",
    "requests>=2.32.3",
    "python-dotenv>=1.1.0",
    "starsessions>=2.2.1",
]
